+++ src/VsDebugMcp.Protocol/Class1.cs
@@ -37,8 +37,10 @@ public static class BridgeMethods
 	public const string DebuggerStart = "debuggerStart";
 	public const string DebuggerEvaluateExpressions = "debuggerEvaluateExpressions";
 	public const string DebuggerGetLocals = "debuggerGetLocals";
+	public const string DebuggerGetThreads = "debuggerGetThreads";
 	public const string TestGet = "testGet";
 	public const string TestRun = "testRun";
+	public const string TestDebug = "testDebug";
 	public const string TestGetStatus = "testGetStatus";
 	public const string TestCancel = "testCancel";
 	public const string Shutdown = "shutdown";
@@ -76,6 +78,7 @@ public static class BridgeErrorCodes
 	public const string DebuggerAlreadyRunning = "debugger_already_running";
 	public const string TestRunBusy = "test_run_busy";
 	public const string TestRunNotFound = "test_run_not_found";
+	public const string TestNotFound = "test_not_found";
 	public const string TestWindowUnavailable = "test_window_unavailable";
 	public const string InternalError = "internal_error";
 }